pyproject.toml
[tool.poetry]
name = "pytest-asyncio-452"
version = "0.1.0"
description = ""
authors = ["Nikolay Bryskin <nbryskin@gmail.com>"]
readme = "README.md"
packages = [{include = "pytest_asyncio_452"}]

[tool.poetry.dependencies]
python = "^3.10"


[tool.poetry.group.dev.dependencies]
anyio = "^3.6.2"
pytest-asyncio = "^0.20.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

